+++ src/sbe_ctd_proc/gui2/__init__.py
@@ -1,7 +1,7 @@
 from nicegui import ui
 # load ui.page routes
 from .ctd_file_page import sbe_plot
-from .app import overview_page
+from .overview_page import overview_page
 
 def start_gui():
     # reload=False avoids error "You must call ui.run() to start the server."
